Refactor repo to become a monorepo
