+++ apps/cli/package.json
@@ -1,0 +1,26 @@
+{
+    "name": "@watonomous/watcloud-email-cli",
+    "bin": {
+        "watcloud-emails": "dist/index.js"
+    },
+    "files": [
+        "dist"
+    ],
+    "scripts": {
+        "dev": "tsc -w",
+        "build": "tsc",
+        "prepack": "npm run build",
+        "check-types": "tsc --noEmit",
+        "clean": "rm -rf dist"
+    },
+    "dependencies": {
+        "@watonomous/watcloud-email-templates": "*",
+        "@repo/utils": "*",
+        "@react-email/components": "0.0.25",
+        "commander": "^12.1.0"
+    },
+    "devDependencies": {
+        "@repo/ts-config": "*",
+        "typescript": "^5.6.3"
+    }
+}
