+++ apps/cli/tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "extends": "@repo/ts-config/tsconfig.json",
+    "compilerOptions": {
+        "rootDir": "src",
+        "outDir": "dist"
+    },
+    "include": [
+        "src"
+    ]
+}
