+++ packages/templates/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@watonomous/watcloud-email-templates",
+  "files": [
+    "dist"
+  ],
+  "exports": {
+    ".": "./dist/index.js",
+    "./*": "./dist/*.js"
+  },
+  "scripts": {
+    "dev": "NODE_ENV=development concurrently \"email dev --dir src/emails\" \"tsc -w\"",
+    "build": "tsc",
+    "prepack": "npm run build",
+    "check-types": "tsc --noEmit",
+    "clean": "rm -rf dist"
+  },
+  "dependencies": {
+    "@react-email/components": "0.0.25",
+    "@repo/utils": "*",
+    "dedent-js": "^1.0.1",
+    "react": "^18.3.1",
+    "react-dom": "18.3.1",
+    "react-email": "3.0.1",
+    "sharp": "^0.34.2",
+    "zod": "^3.23.8"
+  },
+  "devDependencies": {
+    "@repo/ts-config": "*",
+    "@types/react": "^18.3.18",
+    "@types/react-dom": "18.3.0",
+    "concurrently": "^9.2.0",
+    "typescript": "^5.6.3"
+  }
+}
