+++ packages/templates/scripts/perf.sh
