+++ packages/templates/src/emails/blog-update.tsx
@@ -7,7 +7,7 @@ import {
     Text
 } from "@react-email/components";
 import { z } from "zod";
-import { getAsset, registerAsset, WATcloudURI } from "../utils/watcloud-uri";
+import { getAsset, registerAsset, WATcloudURI } from "@repo/utils/watcloud-uri";
 import { WATcloudEmail } from "./_common/watcloud-email";
 
 const WATcloudBlogUpdateEmailProps = z.object({
