+++ packages/templates/src/emails/file-deletion.tsx
