+++ packages/templates/src/emails/mailing-list.tsx
