+++ packages/ts-config/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@repo/ts-config",
+  "description": "Shared TypeScript configuration for the monorepo"
+}
